Break cli into module
